+++ packages/shards-accounts-ui-antd/src/index.ts
@@ -1,0 +1,4 @@
+export { RegistrationFormUI, RegistrationFormUIProps } from './RegistrationFormUi';
+export { ForgotPasswordFormUI, ForgotPasswordFormUIProps } from './ForgotPasswordFormUI';
+export { PasswordResetFormUI, PasswordResetFormUIProps } from './PasswordResetFormUi';
+export { LoginFormUI, LoginFormUIProps } from './LoginFormUi';